Day 7

+++ 2023/rust/day2/src/main.rs
@@ -4,7 +4,7 @@ use common_2023::*;
 mod logic;
 mod tests;
 
-pub const DAY: i32 = 2;
+pub const DAY: u8 = 2;
 
 fn main() {
     run(DAY, "real", part_1, part_2);
