+++ 2023/rust/day3/src/main.rs
@@ -4,7 +4,7 @@ use common_2023::*;
 mod logic;
 mod tests;
 
-pub const DAY: i32 = 3;
+pub const DAY: u8 = 3;
 
 fn main() {
     run(DAY, "real", part_1, part_2);
