+++ 2023/rust/day4/src/main.rs
@@ -4,7 +4,7 @@ use common_2023::*;
 mod logic;
 mod tests;
 
-pub const DAY: i32 = 4;
+pub const DAY: u8 = 4;
 
 fn main() {
     run(DAY, "real", part_1, part_2);
